Crate mrusty

mrusty. mruby safe bindings for Rust

mrusty lets you:

run Ruby 1.9 files with a very restricted API (without having to install Ruby)

reflect Rust structs and enums in mruby and run them

It does all this in a safely neat way while also bringing spec testing and a REPL to the table.

Macros

describe!
A macro useful to run mruby specs. You can pass a tuple of MrubyFiles dependencies as a second argument.

mrclass!
A macro that comes in handy when defining class in order to remove a large part of the clutter and ensure correction. It automates and simplifies the implementation of the MrubyFile trait. Thus, any type provided to mrclass! will get an MrubyFile implementation.

mrfn!
A macro useful for defining Rust closures for mruby. Requires use mrusty::*;.

Structs

GnuReadLine
A struct that implements ReadLine with very basic gnureadline functionality. Requires gnu-readline build feature.

Mruby
A safe struct for the mruby API. The struct only contains creation and desctruction methods. Creating an Mruby returns a MrubyType (Rc<RefCell<Mruby>>) which implements MrubyImpl where the rest of the implemented API is found.

Repl
A struct that exposes an Mruby to a REPL.

Spec
A struct useful for mruby spec definition and running.

Value
A struct that wraps around any mruby variable.

Enums

MrubyError
An enum containing all possbile types of errors.

Traits

MrubyFile
A trait useful for organising Rust types into dynamic mruby files.

MrubyImpl
A trait used on MrubyType which implements mruby functionality.

ReadLine
A trait which performs very basic readline utility.

Type Definitions

MrubyType
A type wrapper around a Rc<RefCell<Mruby>>. Created with Mruby::new().