Struct mrusty::Value

pub struct Value {
    // some fields omitted
}

A struct that wraps around any mruby variable.

Values are created from the MRuby instance:

nil

bool

fixnum

float

string

obj

option

array

Examples

let mruby = MRuby::new();
let result = mruby.run("true").unwrap(); // Value

// Values need to be unwrapped in order to make sure they have the right mruby type.
assert_eq!(result.to_bool().unwrap(), true);

Methods

impl Value

fn init<T: Any>(self, obj: T) -> Value

Initializes the self mruby object passed to initialize with a Rust object of type T.

Note: T must be defined on the current MRuby with def_class.

Examples

use mrusty::*;

let mruby = MRuby::new();

struct Cont {
    value: i32
};

mruby.def_class::<Cont>("Container");
mruby.def_method::<Cont, _>("initialize", mrfn!(|_mruby, slf: Value, v: i32| {
    let cont = Cont { value: v };

    slf.init(cont) // Return the same slf value.
}));

let result = mruby.run("Container.new 3").unwrap();

assert_eq!(result.to_obj::<Cont>().unwrap().value, 3);

fn call(&self, name: &str, args: Vec<Value>) -> Result<Value, MRubyError>

Calls method name on a Value passing args.

Examples

let mruby = MRuby::new();

let one = mruby.fixnum(1);
let result = one.call("+", vec![mruby.fixnum(2)]).unwrap();

assert_eq!(result.to_i32().unwrap(), 3);

fn call_unchecked(&self, name: &str, args: Vec<Value>) -> Value

Calls method name on a Value passing args. If call fails, mruby will be left to handle the exception.

Examples

let mruby = MRuby::new();

let one = mruby.string("");
one.call("+", vec![mruby.fixnum(1)]);

fn to_bool(&self) -> Result<bool, MRubyError>

Casts a Value and returns a bool in an Ok or an Err if the types mismatch.

Example

let mruby = MRuby::new();
let result = mruby.run("
  def pos(n)
    n > 0
  end

  pos 1
").unwrap();

assert_eq!(result.to_bool().unwrap(), true);

fn to_i32(&self) -> Result<i32, MRubyError>

Casts a Value and returns an i32 in an Ok or an Err if the types mismatch.

Example

let mruby = MRuby::new();
let result = mruby.run("
  def fact(n)
    n > 1 ? fact(n - 1) * n : 1
  end

  fact 5
").unwrap();

assert_eq!(result.to_i32().unwrap(), 120);

fn to_f64(&self) -> Result<f64, MRubyError>

Casts a Value and returns an f64 in an Ok or an Err if the types mismatch.

Example

let mruby = MRuby::new();
let result = mruby.run("
  3 / 2.0
").unwrap();

assert_eq!(result.to_f64().unwrap(), 1.5);

fn to_str<'a>(&self) -> Result<&'a str, MRubyError>

Casts a Value and returns a &str in an Ok or an Err if the types mismatch.

Example

let mruby = MRuby::new();
let result = mruby.run("
  [1, 2, 3].map(&:to_s).join
").unwrap();

assert_eq!(result.to_str().unwrap(), "123");

fn to_obj<T: Any>(&self) -> Result<Rc<T>, MRubyError>

Casts mruby Value of Class name to Rust type Rc<T>.

Note: T must be defined on the current MRuby with def_class.

Examples

let mruby = MRuby::new();

struct Cont {
    value: i32
}

mruby.def_class::<Cont>("Container");

let value = mruby.obj(Cont { value: 3 });
let cont = value.to_obj::<Cont>().unwrap();

assert_eq!(cont.value, 3);

fn to_option<T: Any>(&self) -> Result<Option<Rc<T>>, MRubyError>

Casts mruby Value of Class name to Rust Option of Rc<T>.

Note: T must be defined on the current MRuby with def_class.

Examples

let mruby = MRuby::new();

struct Cont {
    value: i32
}

mruby.def_class::<Cont>("Container");

let value = mruby.obj(Cont { value: 3 });
let cont = value.to_option::<Cont>().unwrap();

assert_eq!(cont.unwrap().value, 3);
assert!(mruby.nil().to_option::<Cont>().unwrap().is_none());

fn to_vec(&self) -> Result<Vec<Value>, MRubyError>

Casts mruby Value of Class Array to Rust type Vec<Value>.

Examples

let mruby = MRuby::new();
let result = mruby.run("
  [1, 2, 3].map(&:to_s)
").unwrap();

assert_eq!(result.to_vec().unwrap(), vec![
    mruby.string("1"),
    mruby.string("2"),
    mruby.string("3")
]);