Trait mrusty::MRubyImpl

pub trait MRubyImpl {
    fn filename(&self, filename: &str);
    fn run(&self, script: &str) -> Result<Value, MRubyError>;
    fn runb(&self, script: &[u8]) -> Result<Value, MRubyError>;
    fn execute(&self, script: &Path) -> Result<Value, MRubyError>;
    fn raise(&self, message: &str);
    fn def_file<T: MRubyFile>(&self, name: &str);
    fn def_class<T: Any>(&self, name: &str);
    fn def_method<T: Any, F>(&self, name: &str, method: F) where F: Fn(MRubyType, Value) -> Value + 'static;
    fn def_class_method<T: Any, F>(&self, name: &str, method: F) where F: Fn(MRubyType, Value) -> Value + 'static;
    fn class_name<T: Any>(&self) -> Result<String, MRubyError>;
    fn nil(&self) -> Value;
    fn bool(&self, value: bool) -> Value;
    fn fixnum(&self, value: i32) -> Value;
    fn float(&self, value: f64) -> Value;
    fn string(&self, value: &str) -> Value;
    fn obj<T: Any>(&self, obj: T) -> Value;
    fn option<T: Any>(&self, obj: Option<T>) -> Value;
    fn array(&self, value: Vec<Value>) -> Value;
}

A trait used on MRubyType which implements mruby functionality.

Required Methods

fn filename(&self, filename: &str)

Adds a filename to the mruby context.

Examples

let mruby = MRuby::new();
mruby.filename("script.rb");

let result = mruby.run("1.nope");

match result {
    Err(MRubyError::Runtime(err)) => {
        assert_eq!(err, "script.rb:1: undefined method \'nope\' for 1 (NoMethodError)");
},
    _ => assert!(false)
}

fn run(&self, script: &str) -> Result<Value, MRubyError>

Runs mruby script on a state and context and returns a Value in an Ok or an Err containing an mruby Exception's message.

Examples

let mruby = MRuby::new();
let result = mruby.run("true").unwrap();

assert_eq!(result.to_bool().unwrap(), true);

let mruby = MRuby::new();
let result = mruby.run("'' + 1");

match result {
    Err(MRubyError::Runtime(err)) => {
        assert_eq!(err, "TypeError: expected String");
},
    _ => assert!(false)
}

fn runb(&self, script: &[u8]) -> Result<Value, MRubyError>

Runs mruby compiled (.mrb) script on a state and context and returns a Value in an Ok or an Err containing an mruby Exception's message.

Examples

let mruby = MRuby::new();
let result = mruby.runb(include_bytes!("script.mrb")).unwrap();

fn execute(&self, script: &Path) -> Result<Value, MRubyError>

Runs mruby (compiled (.mrb) or not (.rb)) script on a state and context and returns a Value in an Ok or an Err containing an mruby Exception's message.

Examples

let mruby = MRuby::new();
let result = mruby.execute(File::open("script.rb")).unwrap();

fn raise(&self, message: &str)

Raises an mruby RuntimeError with message message.

Examples

use mrusty::*;

let mruby = MRuby::new();

struct Cont;

mruby.def_class::<Cont>("Container");
mruby.def_class_method::<Cont, _>("hi", mrfn!(|mruby, _slf: Value| {
    mruby.raise("hi");

    mruby.nil()
}));

let result = mruby.run("Container.hi");

match result {
    Err(MRubyError::Runtime(err)) => {
        assert_eq!(err, "RuntimeError: hi");
},
    _ => assert!(false)
}

fn def_file<T: MRubyFile>(&self, name: &str)

Defines a dynamic file that can be required containing the Rust type T and runs its MRubyFile-inherited require method.

Examples

use mrusty::*;

let mruby = MRuby::new();

struct Cont {
    value: i32
};

impl MRubyFile for Cont {
    fn require(mruby: MRubyType) {
        mruby.def_class::<Cont>("Container");
        mruby.def_method::<Cont, _>("initialize", mrfn!(|_mruby, slf: Value, v: i32| {
            let cont = Cont { value: v };

            slf.init(cont)
        }));
        mruby.def_method::<Cont, _>("value", mrfn!(|mruby, slf: Cont| {
            mruby.fixnum(slf.value)
        }));
    }
}

mruby.def_file::<Cont>("cont");

let result = mruby.run("
    require 'cont'

    Container.new(3).value
").unwrap();

assert_eq!(result.to_i32().unwrap(), 3);

fn def_class<T: Any>(&self, name: &str)

Defines Rust type T as an mruby Class named name.

Examples

let mruby = MRuby::new();

struct Cont {
    value: i32
}

mruby.def_class::<Cont>("Container");

fn def_method<T: Any, F>(&self, name: &str, method: F) where F: Fn(MRubyType, Value) -> Value + 'static

Defines an mruby method named name. The closure to be run when the name method is called should be passed through the mrfn! macro.

Examples

use mrusty::*;

let mruby = MRuby::new();

struct Cont {
    value: i32
};

mruby.def_class::<Cont>("Container");
mruby.def_method::<Cont, _>("initialize", mrfn!(|_mruby, slf: Value, v: i32| {
    let cont = Cont { value: v };

    slf.init(cont)
}));
mruby.def_method::<Cont, _>("value", mrfn!(|mruby, slf: Cont| {
    mruby.fixnum(slf.value)
}));

let result = mruby.run("Container.new(3).value").unwrap();

assert_eq!(result.to_i32().unwrap(), 3);

fn def_class_method<T: Any, F>(&self, name: &str, method: F) where F: Fn(MRubyType, Value) -> Value + 'static

Defines an mruby class method named name. The closure to be run when the name method is called should be passed through the mrfn! macro.

Examples

use mrusty::*;

let mruby = MRuby::new();

struct Cont;

mruby.def_class::<Cont>("Container");
mruby.def_class_method::<Cont, _>("hi", mrfn!(|mruby, _slf: Value, v: i32| {
    mruby.fixnum(v)
}));

let result = mruby.run("Container.hi 3").unwrap();

assert_eq!(result.to_i32().unwrap(), 3);

fn class_name<T: Any>(&self) -> Result<String, MRubyError>

Return the mruby name of a previously defined Rust type T with def_class.

Examples

let mruby = MRuby::new();

struct Cont;

mruby.def_class::<Cont>("Container");

assert_eq!(mruby.class_name::<Cont>().unwrap(), "Container");

fn nil(&self) -> Value

Creates mruby Value nil.

Examples

let mruby = MRuby::new();

struct Cont;

mruby.def_class::<Cont>("Container");
mruby.def_method::<Cont, _>("nil", |mruby, _slf| mruby.nil());

let result = mruby.run("Container.new.nil.nil?").unwrap();

assert_eq!(result.to_bool().unwrap(), true);

fn bool(&self, value: bool) -> Value

Creates mruby Value containing true or false.

Examples

let mruby = MRuby::new();

let b = mruby.bool(true);

assert_eq!(b.to_bool().unwrap(), true);

fn fixnum(&self, value: i32) -> Value

Creates mruby Value of Class Fixnum.

Examples

let mruby = MRuby::new();

let fixn = mruby.fixnum(2);

assert_eq!(fixn.to_i32().unwrap(), 2);

fn float(&self, value: f64) -> Value

Creates mruby Value of Class Float.

Examples

let mruby = MRuby::new();

let fl = mruby.float(2.3);

assert_eq!(fl.to_f64().unwrap(), 2.3);

fn string(&self, value: &str) -> Value

Creates mruby Value of Class String.

Examples

let mruby = MRuby::new();

let s = mruby.string("hi");

assert_eq!(s.to_str().unwrap(), "hi");

fn obj<T: Any>(&self, obj: T) -> Value

Creates mruby Value of Class name containing a Rust object of type T.

Note: T must be defined on the current MRuby with def_class.

Examples

let mruby = MRuby::new();

struct Cont {
    value: i32
}

mruby.def_class::<Cont>("Container");

let value = mruby.obj(Cont { value: 3 });

fn option<T: Any>(&self, obj: Option<T>) -> Value

Creates mruby Value of Class name containing a Rust Option of type T.

Note: T must be defined on the current MRuby with def_class.

Examples

let mruby = MRuby::new();

struct Cont {
    value: i32
}

mruby.def_class::<Cont>("Container");

let none = mruby.option::<Cont>(None);
let some = mruby.option(Some(Cont { value: 3 }));

assert_eq!(none.call("nil?", vec![]).unwrap().to_bool().unwrap(), true);
assert_eq!(some.to_obj::<Cont>().unwrap().value, 3);

fn array(&self, value: Vec<Value>) -> Value

Creates mruby Value of Class Array.

Examples

let mruby = MRuby::new();

let array = mruby.array(vec![
    mruby.fixnum(1),
    mruby.fixnum(2),
    mruby.fixnum(3)
]);

assert_eq!(array.to_vec().unwrap(), vec![
    mruby.fixnum(1),
    mruby.fixnum(2),
    mruby.fixnum(3)
]);