Crate mrusty

mrusty. mruby safe bindings for Rust

Requirements

mruby

mrusty requires mruby compiled with fPIC. To compile and install mruby 1.2.0:

make sure you have Bison & Ruby installed

download the source

unzip and cd to mruby-1.2.0/

add the following lines to build_config.rb as in the # C compiler settings example:

conf.cc do |cc|
    cc.flags << '-fPIC'
end

run ./minirake

Macros

describe!
A macro useful to run mruby specs.

mrfn!
A macro useful for defining Rust closures for mruby. Requires use mrusty::*;.

Structs

MRuby
A safe struct for the mruby API. The struct only contains creation and desctruction methods. Creating an MRuby returns a MRubyType (Rc<RefCell<MRuby>>) which implements MRubyImpl where the rest of the implemented API is found.

Spec
A struct useful for mruby spec definition and running.

Value
A struct that wraps around any mruby variable.

Enums

MRubyError
An enum containing all possbile types of errors.

Traits

MRubyFile
A trait useful for organising Rust types into dynamic mruby files.

MRubyImpl
A trait used on MRubyType which implements mruby functionality.

Type Definitions

MRubyType
A type wrapper around a Rc<RefCell<MRuby>>. Created with MRuby::new().